--- a/Project.toml
+++ b/Project.toml
@@ -5,6 +5,7 @@ version = "0.4.5"
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10, 1.0"