--- a/Project.toml
+++ b/Project.toml
@@ -9,6 +9,7 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10, 1.0"
+StaticArrays = "1.0"
 julia = "1"
 
 [extras]