--- a/src/FastGaussQuadrature.jl
+++ b/src/FastGaussQuadrature.jl
@@ -1,12 +1,7 @@
 module FastGaussQuadrature
 
-using SpecialFunctions, LinearAlgebra
-cumprod(A::AbstractArray) = Base.cumprod(A, dims=1)
-cumprod(A::AbstractArray, d::Int) = Base.cumprod(A, dims=d)
-sum(A::AbstractArray, n::Int) = Base.sum(A, dims=n)
-sum(A) = Base.sum(A)
-flipdim(A, d) = reverse(A, dims=d)
-
+using LinearAlgebra
+using SpecialFunctions
 
 export gausslegendre
 export gausschebyshev