--- a/src/besselroots.jl
+++ b/src/besselroots.jl
@@ -162,3 +162,88 @@ function Piessens(ν::Float64)
     y[1] *= sqrt(ν+1)  # Scale the first root.
     return y
 end
+
+
+function besselZeroRoots(m)
+    # BESSEL0ROOTS ROOTS OF BESSELJ(0,x). USE ASYMPTOTICS.
+    # Use McMahon's expansion for the remainder (NIST, 10.21.19):
+    jk = Array{Float64}(undef, m)
+    p = (1071187749376 / 315, 0.0, -401743168 / 105, 0.0, 120928 / 15,
+         0.0, -124 / 3, 0.0, 1.0, 0.0)
+    # First 20 are precomputed:
+    @inbounds for jj = 1:min(m, 20)
+        jk[jj] = J0_roots[jj]
+    end
+    @inbounds for jj = 21:min(m, 47)
+        ak = π * (jj - .25)
+        ak82 = (.125 / ak)^2
+        jk[jj] = ak + .125 / ak * @evalpoly(ak82, 1.0, p[7], p[5], p[3])
+    end
+    @inbounds for jj = 48:min(m, 344)
+        ak = π * (jj - .25)
+        ak82 = (.125 / ak)^2
+        jk[jj] = ak + .125 / ak * @evalpoly(ak82, 1.0, p[7], p[5])
+    end
+    @inbounds for jj = 345:min(m,13191)
+        ak = π * (jj - .25)
+        ak82 = (.125 / ak)^2
+        jk[jj] = ak + .125 / ak * muladd(ak82, p[7], 1.0)
+    end
+    @inbounds for jj = 13192:m
+        ak = π * (jj - .25)
+        jk[jj] = ak + .125 / ak
+    end
+    return jk
+end
+
+const besselJ1_10 = [0.2695141239419169, 0.1157801385822037,
+                     0.07368635113640822, 0.05403757319811628,
+                     0.04266142901724309, 0.03524210349099610,
+                     0.03002107010305467, 0.02614739149530809,
+                     0.02315912182469139, 0.02078382912226786]
+
+function besselJ1(m)
+    # BESSELJ1 EVALUATE BESSELJ(1,x)^2 AT ROOTS OF BESSELJ(0,x).
+    # USE ASYMPTOTICS. Use Taylor series of (NIST, 10.17.3) and McMahon's
+    # expansion (NIST, 10.21.19):
+    Jk2 = Array{Float64}(undef, m)
+    c = (-171497088497 / 15206400, 461797 / 1152, -172913 / 8064,
+         151 / 80, -7 / 24, 0.0, 2.0)
+    # First 10 are precomputed:
+    @inbounds for jj = 1:min(m, 10)
+        Jk2[jj] = besselJ1_10[jj]
+    end
+    @inbounds for jj = 11:min(m, 15)
+        ak = π * (jj - .25)
+        ak2 = (1 / ak)^2
+        Jk2[jj] = 1 / (π * ak) * muladd(@evalpoly(ak2, c[5], c[4], c[3],
+                                                  c[2], c[1]), ak2^2, c[7])
+    end
+    @inbounds for jj = 16:min(m, 21)
+        ak = π * (jj - .25)
+        ak2 = (1 / ak)^2
+        Jk2[jj] = 1 / (π * ak) * muladd(@evalpoly(ak2, c[5], c[4], c[3], c[2]),
+                                        ak2^2, c[7])
+    end
+    @inbounds for jj = 22:min(m,55)
+        ak = π * (jj - .25)
+        ak2 = (1 / ak)^2
+        Jk2[jj] = 1 / (π * ak) * muladd(@evalpoly(ak2, c[5], c[4], c[3]),
+                                        ak2^2, c[7])
+    end
+    @inbounds for jj = 56:min(m,279)
+        ak = π * (jj - .25)
+        ak2 = (1 / ak)^2
+        Jk2[jj] = 1 / (π * ak) * muladd(muladd(ak2, c[4], c[5]), ak2^2, c[7])
+    end
+    @inbounds for jj = 280:min(m,2279)
+        ak = π * (jj - .25)
+        ak2 = (1 / ak)^2
+        Jk2[jj] = 1 / (π * ak) * muladd(ak2^2, c[5], c[7])
+    end
+    @inbounds for jj = 2280:m
+        ak = π * (jj - .25)
+        Jk2[jj] = 1 / (π * ak) * c[7]
+    end
+    return Jk2
+end