--- a/src/gaussjacobi.jl
+++ b/src/gaussjacobi.jl
@@ -369,12 +369,12 @@ function feval_asy1(n::Integer, α::Float64, β::Float64, t::AbstractVector, idx
     return vals, ders
 end
 
-function boundary(n::Integer, α::Float64, β::Float64, npts)
+function boundary(n::Integer, α::Float64, β::Float64, npts::Integer)
 # Algorithm for computing nodes and weights near the boundary.
 
     # Use Newton iterations to find the first few Bessel roots:
     smallK = min(30, npts)
-    jk = besselroots(α, min(npts, smallK))
+    jk = besselroots(α, smallK)
 
     # Approximate roots via asymptotic formula: (see Olver 1974)
     phik = jk/(n + .5*(α + β + 1))