# Improvements
## Faster `besselroots`
### Case 1 (33% faster!)
current master:
```julia
julia> @benchmark besselroots(0.3, 4)
BenchmarkTools.Trial: 
  memory estimate:  576 bytes
  allocs estimate:  3
  --------------
  minimum time:     201.992 ns (0.00% GC)
  median time:      209.326 ns (0.00% GC)
  mean time:        235.806 ns (9.05% GC)
  maximum time:     3.493 μs (93.49% GC)
  --------------
  samples:          10000
  evals/sample:     631
```

with this PR:
```julia
julia> @benchmark besselroots(0.3, 4)
BenchmarkTools.Trial: 
  memory estimate:  240 bytes
  allocs estimate:  2
  --------------
  minimum time:     143.281 ns (0.00% GC)
  median time:      146.550 ns (0.00% GC)
  mean time:        157.969 ns (3.62% GC)
  maximum time:     1.952 μs (89.20% GC)
  --------------
  samples:          10000
  evals/sample:     846
```

### Case 2 (17% faster!)
current master:
```julia
julia> @benchmark besselroots(0.3, 10)
BenchmarkTools.Trial: 
  memory estimate:  624 bytes
  allocs estimate:  3
  --------------
  minimum time:     322.403 ns (0.00% GC)
  median time:      333.927 ns (0.00% GC)
  mean time:        369.698 ns (5.34% GC)
  maximum time:     8.332 μs (95.34% GC)
  --------------
  samples:          10000
  evals/sample:     233
```

with this PR:

```julia
julia> @benchmark besselroots(0.3, 10)
BenchmarkTools.Trial: 
  memory estimate:  288 bytes
  allocs estimate:  2
  --------------
  minimum time:     266.314 ns (0.00% GC)
  median time:      274.412 ns (0.00% GC)
  mean time:        305.570 ns (1.70% GC)
  maximum time:     3.610 μs (88.85% GC)
  --------------
  samples:          10000
  evals/sample:     318
```

### Notice
But, this improvement is not much efficient to `gauss*` functions.
Because `besselroots` function is not their bottleneck.

## Faster `gausslegendre`
### Case 1 (15% faster!)
current master:
```julia
julia> @benchmark gausslegendre(5)
BenchmarkTools.Trial: 
  memory estimate:  256 bytes
  allocs estimate:  2
  --------------
  minimum time:     51.098 ns (0.00% GC)
  median time:      55.798 ns (0.00% GC)
  mean time:        67.776 ns (12.27% GC)
  maximum time:     2.157 μs (96.36% GC)
  --------------
  samples:          10000
  evals/sample:     987
```

with this PR:
```julia
julia> @benchmark gausslegendre(5)
BenchmarkTools.Trial: 
  memory estimate:  256 bytes
  allocs estimate:  2
  --------------
  minimum time:     41.976 ns (0.00% GC)
  median time:      48.264 ns (0.00% GC)
  mean time:        57.450 ns (8.62% GC)
  maximum time:     1.408 μs (95.70% GC)
  --------------
  samples:          10000
  evals/sample:     991
```

### Notice
This improvement is efficient only for small `n`.

## Solve type unstability
current master:
![image](https://user-images.githubusercontent.com/7488140/104815853-a694c280-585a-11eb-8539-afd9209d5619.png)

with this PR:
![image](https://user-images.githubusercontent.com/7488140/104815864-b3b1b180-585a-11eb-992b-2ef3fcdb0a2a.png)