@dlfivefifty 2021-01-16
It looks fine to me. Please bump the patch version and tag a new release (using `@JuliaRegistrator register`)