@JuliaRegistrator 2021-01-16
Comments on pull requests will not trigger Registrator, as it is disabled. Please try commenting on a commit or issue.